Still structing data

+++ projeto-cidades-integra-back/src/main/java/com/jamalxvi/projetocidadesintegra/projetocidadesintegra/models/City.java
@@ -4,6 +4,7 @@ import lombok.Builder;
 import lombok.Data;
 
 import javax.persistence.*;
+import java.math.BigDecimal;
 
 @Entity
 @Table(name="CITY")
@@ -21,9 +22,9 @@ public class City {
     @Column(name = "CAPITAL")
     private Boolean capital;
     @Column(name = "LONG")
-    private Long longitude;
+    private BigDecimal longitude;
     @Column(name = "LAT")
-    private Long latitude;
+    private BigDecimal latitude;
     @Column(name = "NO_ACCENTS")
     private String nameWithOutAccent;
     @Column(name = "ALTERNATIVE_NAMES")
